fetch/api/request/destination/fetch-destination-fetch.https.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="../../../../service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script>
async_test(function(t) {
    var worker_url = 'resources/fetch-destination-worker.js';
    var scope = 'resources/empty.html';
    var registration;

    window.onmessage = t.step_func(function(e) {
        assert_true(e.data);
        t.done();
    });
    service_worker_unregister_and_register(t, worker_url, scope)
      .then(t.step_func(function(r) {
          registration = r;
          return wait_for_state(t, r.installing, 'activated');
        }))
      .then(t.step_func(function() {
          return with_iframe(scope);
        }))
      .then(t.step_func(function(frame) {
          frame.contentWindow.registration = registration;
          var node = frame.contentWindow.document.createElement("script");
          var text = frame.contentWindow.document.createTextNode(
              'fetch("dummy?dest=").then(function(response) { ' +
                  'registration.unregister();' +
                  'var url = new URL(parent.location);' +
                  'parent.postMessage(response.ok, url.origin); })');
          node.appendChild(text);
          frame.contentWindow.document.body.appendChild(node);
        }))
      .catch(unreached_rejection(t));
}, 'Fetch destination fetch');
</script>
